# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stager API
  description: |
    API for Stager - AI-Powered Virtual Staging Platform for Real Estate.

    ## Authentication
    All endpoints (except webhooks) require authentication via Supabase Auth session cookies.
    Unauthenticated requests return `401 Unauthorized`.

    ## Rate Limiting
    Resource-intensive endpoints are rate-limited:
    - **Staging:** 20 requests/minute
    - **Email:** 10 requests/minute
    - **Auth:** 5 requests/minute

    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Requests remaining in window
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Error Responses
    All errors follow a consistent format with actionable suggestions:
    ```json
    {
      "error": "Human-readable message",
      "code": "ERROR_CODE",
      "suggestion": "Actionable suggestion for the user",
      "action": { "label": "Button label", "href": "/path" },
      "details": { "field": "value" }
    }
    ```

    ## Credits
    - Staging costs 1 credit per image
    - First 2 remixes per original image are free
    - Additional remixes cost 1 credit each
  version: 1.136.0
  contact:
    name: Stager Support
servers:
  - url: https://stager-tau.vercel.app/api
    description: Production
  - url: http://localhost:3000/api
    description: Local Development
...
paths:
...
  /webhooks/stripe:
    post:
      tags: [Webhooks]
      summary: Stripe webhook
      description: |
        Receives Stripe webhook events for subscription and payment updates.

        **Authentication:** Validates Stripe signature header (no user auth required).

        Handled events:
        - `checkout.session.completed`: Process new subscription or top-up
        - `customer.subscription.updated`: Update subscription status
        - `customer.subscription.deleted`: Handle subscription cancellation
        - `invoice.payment_succeeded`: Credit renewal
        - `invoice.payment_failed`: Handle failed payments
      operationId: stripeWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed
        '400':
          description: Invalid signature or payload
  /webhooks/replicate:
    post:
      tags: [Webhooks]
      summary: Replicate webhook
      description: |
        Receives Replicate webhook callbacks when async staging jobs complete.

        **Note:** Currently does not validate webhook signature (improvement needed).
      operationId: replicateWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: Prediction ID
                status:
                  type: string
                  enum: [succeeded, failed, canceled]
                output:
                  oneOf:
                    - type: string
                    - type: array
                      items:
                        type: string
                  description: Output image URL(s)
                error:
                  type: string
                  description: Error message if failed
      responses:
        '200':
          description: Webhook processed
...