# src/app/api/webhooks/replicate/route.ts
import { NextRequest, NextResponse } from "next/server";
import { createClient } from "@supabase/supabase-js";
import { CREDITS_PER_STAGING, ROOM_TYPES } from "@/lib/constants";
import { createNotification } from "@/lib/notifications";
import { validateReplicateWebhook } from "@/lib/webhooks/validation";
import { deductCredits } from "@/lib/billing/credits.service";
import { getRequestId, REQUEST_ID_HEADER } from "@/lib/api/request-id";

/**
 * Replicate webhook payload structure
 */
interface ReplicateWebhookPayload {
  id: string;
  status: "starting" | "processing" | "succeeded" | "failed" | "canceled";
  output?: string[] | null;
  error?: string | null;
  metrics?: {
    predict_time?: number;
  };
}

/**
 * POST /api/webhooks/replicate
 *
 * Receives completion callbacks from Replicate API.
 * Updates staging job status and downloads generated image.
 */
export async function POST(request: NextRequest) {
  const startTime = Date.now();
  const requestId = getRequestId(request);

  // Get raw body for signature verification
  const rawBody = await request.text();
  const signature = request.headers.get("webhook-signature");

  // Verify webhook signature
  const validation = validateReplicateWebhook(
    rawBody,
    signature,
    process.env.REPLICATE_WEBHOOK_SECRET
  );

  if (!validation.valid) {
    console.error(`[${requestId}] Invalid webhook signature:`, validation.error);
    const response = NextResponse.json({ error: "Invalid signature" }, { status: 401 });
    response.headers.set(REQUEST_ID_HEADER, requestId);
    return response;
  }

  // Parse payload
  let payload: ReplicateWebhookPayload;
  try {
    payload = JSON.parse(rawBody);
  } catch {
    return NextResponse.json({ error: "Invalid JSON" }, { status: 400 });
  }

  // Use service role client for webhook processing
  const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL;
  const supabaseServiceKey = process.env.SUPABASE_SERVICE_ROLE_KEY;

  if (!supabaseUrl || !supabaseServiceKey) {
    console.error("Missing Supabase configuration");
    return NextResponse.json({ error: "Server configuration error" }, { status: 500 });
  }

  const supabase = createClient(supabaseUrl, supabaseServiceKey);

  // Find the staging job by prediction ID
  const { data: job, error: findError } = await supabase
    .from("staging_jobs")
    .select("*")
    .eq("replicate_prediction_id", payload.id)
    .single();

  if (findError || !job) {
    console.error(`Job not found for prediction ${payload.id}:`, findError);
    // Return 200 to prevent retries for unknown predictions
    return NextResponse.json({ message: "Job not found, ignoring" }, { status: 200 });
  }

  // Handle different statuses
  if (payload.status === "succeeded" && payload.output && payload.output.length > 0) {
    // Success - download and store the generated image
    try {
      const outputUrl = payload.output[0];
      const processingTimeMs = payload.metrics?.predict_time
        ? Math.round(payload.metrics.predict_time * 1000)
        : Date.now() - new Date(job.created_at).getTime();

      // Download the generated image
      const imageResponse = await fetch(outputUrl);
      if (!imageResponse.ok) {
        throw new Error(`Failed to download image: ${imageResponse.status}`);
      }

      const imageBuffer = await imageResponse.arrayBuffer();
      const mimeType = imageResponse.headers.get("content-type") || "image/png";
      const ext = mimeType.split("/")[1] || "png";

      // Upload to Supabase Storage
      const stagedFileName = `${job.user_id}/${job.id}-staged.${ext}`;
      const { error: uploadError } = await supabase.storage
        .from("staging-images")
        .upload(stagedFileName, imageBuffer, {
          contentType: mimeType,
          upsert: true,
        });

      if (uploadError) {
        console.error("Failed to upload staged image:", uploadError);
        throw uploadError;
      }

      // Get public URL
      const { data: publicUrlData } = supabase.storage
        .from("staging-images")
        .getPublicUrl(stagedFileName);

      // Update job as completed
      await supabase
        .from("staging_jobs")
        .update({
          status: "completed",
          staged_image_url: publicUrlData.publicUrl,
          completed_at: new Date().toISOString(),
          processing_time_ms: processingTimeMs,
        })
        .eq("id", job.id);

      // Deduct credits using centralized service
      const creditResult = await deductCredits(supabase, job.user_id, CREDITS_PER_STAGING);
      const newCredits = creditResult.newBalance;

      // Create staging completion notification
      const roomLabel = ROOM_TYPES.find((r) => r.id === job.room_type)?.label || job.room_type;
      await createNotification(
        supabase,
        job.user_id,
        "staging_complete",
        "Staging Complete",
        `Your ${roomLabel} staging is ready to view!`,
        "/history"
      );

      // Check for low credits warning (3 or fewer)
      if (newCredits <= 3 && newCredits > 0) {
        await createNotification(
          supabase,
          job.user_id,
          "low_credits",
          "Low Credits",
          `You have ${newCredits} credit${newCredits !== 1 ? "s" : ""} remaining. Consider adding more to continue staging.`,
          "/billing"
        );
      }

    } catch (error) {
      console.error(`Failed to process successful prediction ${payload.id}:`, error);

      // Mark as failed
      await supabase
        .from("staging_jobs")
        .update({
          status: "failed",
          error_message: error instanceof Error ? error.message : "Failed to process result",
          completed_at: new Date().toISOString(),
          processing_time_ms: Date.now() - startTime,
        })
        .eq("id", job.id);
    }
  } else if (payload.status === "failed" || payload.status === "canceled") {
    // Failed - update job status
    await supabase
      .from("staging_jobs")
      .update({
        status: "failed",
        error_message: payload.error || `Prediction ${payload.status}`,
        completed_at: new Date().toISOString(),
        processing_time_ms: Date.now() - new Date(job.created_at).getTime(),
      })
      .eq("id", job.id);

    // Create staging failure notification
    const roomLabel = ROOM_TYPES.find((r) => r.id === job.room_type)?.label || job.room_type;
    await createNotification(
      supabase,
      job.user_id,
      "staging_failed",
      "Staging Failed",
      `Your ${roomLabel} staging could not be completed. Please try again.`,
      "/history"
    );
  }
  // Ignore "starting" and "processing" statuses - they're intermediate

  const response = NextResponse.json({ success: true });
  response.headers.set(REQUEST_ID_HEADER, requestId);
  return response;
}